requirements.txt
fire==0.6.0
numpy==1.24.4
opencv_python==4.8.0.76
Pillow==10.0.0
torch==2.4.0
torchvision==0.19.0
